Hi, I'm PsalmEquipt
For several years, I've trusted God in matters big and small, and He has never let me down. He's in the miracle business every day and will not do less than His Word. Perhaps like Moses, you'd like to find out more.

And the angel of the LORD appeared unto him in a flame of fire out of the midst of a bush: and he looked, and, behold, the bush burned with fire, and the bush was not consumed. And Moses said, I will now turn aside, and see this great sight, why the bush is not burnt. Exodus 3:2,3

Just as Moses did, we can take the time and effort to seek the things of God in our daily lives. Ephesians reveals a glimpse of what's ahead for us as we are faithful in this:

That in the ages to come he might shew the exceeding riches of his grace in his kindness toward us through Christ Jesus. Ephesians 2:7

For a wealth of enlightenment and answers to life's questions visit:

THE WAY INTERNATIONAL Biblical Research, Teaching, and Fellowship Ministry

Founded 1942

Website www.theway.org

More Info psalmequipt.nichesite.org